ai-doc-agent/requirements.txt
# Core LangChain
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>=1.0.0

# Vector Store
chromadb>=0.5.0

# Document Loaders
pypdf>=3.17.0
wikipedia>=1.4.0

# Web Fetching
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Google Gemini
google-generativeai>=0.7.0

# Utilities
python-dotenv>=1.0.0
